requirements.txt
colorama==0.4.6
discord.py==2.5.2
Flask==3.0.3
httpx==0.27.2
requests==2.32.3
tls-client==1.0.1
typing_extensions==4.12.2
pymongo==4.8.0
fastapi==0.115.0
websocket-client==1.8.0
uvicorn[standard]==0.30.6
websockets==13.0.1